app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
prometheus-client>=0.20.0
